Fix type casting for latitude and longitude in RecommendationService to ensure proper distance calculation.

+++ backend/place-recommendation-service/src/main/java/com/meetmate/place/service/RecommendationService.java
@@ -113,7 +113,7 @@ public class RecommendationService {
                 continue;
             }
             double dist = DistanceCalculator.calculateDistance(
-                pref.currentLat(), pref.currentLng(),
+                pref.currentLat().doubleValue(), pref.currentLng().doubleValue(),
                 place.getLat().doubleValue(), place.getLng().doubleValue()
             );
             distances.add(dist);
